[Silver III] Title: 주유소, Time: 20 ms, Memory: 2804 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -20,7 +20,7 @@ int main(){
     for(int i = 0; i < n - 1; i++){
         min_cost = min(min_cost, costs[i]);
 
-        answer += min_cost * roads[i];
+        answer += (long long)min_cost * roads[i];
     }
 
     cout << answer;
